Reorganize project
Consolidates code in lib dir
