+++ lib/plugin.js
