+++ lib/plugin.test.js
@@ -7,7 +7,7 @@ const vfs = require('vinyl-fs');
 const Vinyl = require('gulp-util').File;
 
 const lab = exports.lab = Lab.script();
-const markdown = require('../index');
+const markdown = require('./plugin');
 
 function commonmarkRender (string) {
   const commonmark = require('commonmark');
@@ -44,7 +44,7 @@ const testConfigs = {
   }
 };
 
-const fixturePath = './test/fixtures/**/*';
+const fixturePath = Path.join(process.cwd(), 'test_fixtures', '**', '*');
 
 const fixtureConfig = {
   frontmatter: {
