+++ test_fixtures/blog/posts/bushwick-artisan.md
