+++ test_fixtures/blog/posts/oakland-activist.md
